Fixed problem with IpAddress convertation to integer represantation.

+++ IpToCountry/IpAddressExtensions.cs
@@ -4,12 +4,12 @@ namespace IpToCountry
 {
     public static class IpAddressExtensions
     {
-        public static int ToInteger(this IPAddress ipAddress)
+        public static uint ToInteger(this IPAddress ipAddress)
         {
             var bytes = ipAddress.GetAddressBytes();
             if (bytes.Any())
             {
-                return bytes[0] << 24 | bytes[1] << 16 | bytes[2] << 8 | bytes[3];
+                return (uint)(bytes[0] << 24 | bytes[1] << 16 | bytes[2] << 8 | bytes[3]);
             }
             return 0;
         }
